+++ pj_plugins/dialog_protocol/CMakeLists.txt
@@ -1,9 +1,5 @@
 cmake_minimum_required(VERSION 3.20)
 project(pj_dialog_protocol LANGUAGES CXX)
-set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-set(CMAKE_CXX_EXTENSIONS OFF)
-
 
 # --- Header-only interface libraries ---
 
@@ -19,6 +15,7 @@ target_link_libraries(pj_dialog_protocol INTERFACE pj_base)
 find_package(nlohmann_json REQUIRED)
 
 add_library(pj_dialog_sdk INTERFACE)
+target_compile_features(pj_dialog_sdk INTERFACE cxx_std_17)
 target_link_libraries(pj_dialog_sdk INTERFACE pj_dialog_protocol nlohmann_json::nlohmann_json)
 target_include_directories(pj_dialog_sdk INTERFACE
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -28,6 +25,7 @@ set_target_properties(pj_dialog_sdk PROPERTIES EXPORT_NAME Dialog)
 
 # Host-side C++ API (depends on protocol header + nlohmann/json)
 add_library(pj_dialog_host INTERFACE)
+target_compile_features(pj_dialog_host INTERFACE cxx_std_17)
 target_link_libraries(pj_dialog_host INTERFACE pj_dialog_protocol nlohmann_json::nlohmann_json)
 target_include_directories(pj_dialog_host INTERFACE include)
